src/detect/requirements.txt
pandas==2.0.3
evidently==0.4.0
hydra-core==1.3.2
kestra==0.6.0